requirements.txt
PyQt6==6.6.1
requests==2.31.0
cryptography==41.0.7
Pillow==10.2.0
argon2-cffi==23.1.0